CC-4498: INT to FLOAT refactoring.

+++ Bundles/ProductQuantityStorage/src/Spryker/Client/ProductQuantityStorage/Dependency/Service/ProductQuantityStorageToProductQuantityServiceBridge.php
@@ -26,11 +26,11 @@ class ProductQuantityStorageToProductQuantityServiceBridge implements ProductQua
 
     /**
      * @param \Generated\Shared\Transfer\ProductQuantityTransfer $productQuantityTransfer
-     * @param int $quantity
+     * @param float $quantity
      *
-     * @return int
+     * @return float
      */
-    public function getNearestQuantity(ProductQuantityTransfer $productQuantityTransfer, int $quantity): int
+    public function getNearestQuantity(ProductQuantityTransfer $productQuantityTransfer, float $quantity): float
     {
         return $this->productQuantityService
             ->getNearestQuantity($productQuantityTransfer, $quantity);
